Nose

+++ Aplicacion/src/main.js
@@ -1,12 +1,22 @@
 const { BrowserWindow, Notification } = require('electron')
 const { getConnection } = require('./database')
 
-
-
 async function borrar_venta_main(idVenta) {
     const conn = await getConnection();
     await conn.query('delete from venta_mascota where id_venta = ?', idVenta);
     await conn.query('delete from venta where id_venta = ?', idVenta);
+}
+
+async function restarPuntosClientePorBorradoDeVenta(idVenta, idCliente) {
+    const conn = await getConnection();
+
+    let puntosParaRestar = await conn.query('select puntos_obtenidos from venta where id_venta = ?', idVenta);
+    let puntosActuales = await conn.query('select puntos from clientes where id_cliente = ?', idCliente);
+
+    puntosParaRestar = puntosParaRestar[0].puntos_obtenidos;
+    puntosActuales = puntosActuales[0].puntos;
+
+    await conn.query('UPDATE clientes SET puntos = (? - ?) WHERE id_cliente = ?;', [puntosActuales, puntosParaRestar, idCliente]);
 }
 
 
@@ -439,7 +449,16 @@ async function insertCliente(cliente) {
 
 async function insertVentamain(venta, mascotas) {
     const conn = await getConnection();
-    const resultado3 = await conn.query('update venta set venta.activo = false where venta.id_cliente = ?', venta.id_cliente);
+
+    console.log("Mascotas seleccionadas: ", mascotas);
+    for (let i = 0; i < mascotas.length; i++) {
+        let idVentaActivaMascota = await conn.query('select venta.id_venta from venta inner join venta_mascota on venta.id_venta = venta_mascota.id_venta where venta_mascota.id_mascota = ? and venta.activo = true', mascotas[i]);
+        console.log("idVentaActivaMAscota: ", idVentaActivaMascota);
+        if (idVentaActivaMascota.length !== 0) {
+            await conn.query('update venta set venta.activo = false where venta.id_venta = ?', idVentaActivaMascota[0].id_venta);
+        }
+    }
+
     const resultado = await conn.query('insert into venta set ?', venta);
     const idVenta = await conn.query('select id_venta from venta order by id_venta DESC LIMIT 1');
 
@@ -452,6 +471,8 @@ async function insertVentamain(venta, mascotas) {
     }
 
 }
+
+
 
 
 
@@ -516,7 +537,7 @@ async function get20VentasClientemain(busqueda, salto) {
     } else {
         let busquedaMod = "%" + busqueda + "%";
         console.log("entroo", busquedaMod);
-        const resultado = await conn.query(`select * from venta inner join clientes on clientes.id_cliente = venta.id_cliente inner join bolsas_kilos on bolsas_kilos.id_bolsa_kilo = venta.id_bolsa_kilo inner join bolsas on bolsas.id_bolsa = bolsas_kilos.id_bolsa where venta.precio like ? or venta.totalventa like ? or bolsas.marca_bolsa like ? or bolsas_kilos.kilos_bolsa = ? order by venta.fecha DESC LIMIT ?, 20;`, [busquedaMod, busquedaMod, busquedaMod, parseFloat(busqueda), salto]);
+        const resultado = await conn.query(`select * from venta inner join clientes on clientes.id_cliente = venta.id_cliente inner join bolsas_kilos on bolsas_kilos.id_bolsa_kilo = venta.id_bolsa_kilo inner join bolsas on bolsas.id_bolsa = bolsas_kilos.id_bolsa where venta.precio like ? or venta.totalventa like ? or bolsas.marca_bolsa like ? or bolsas_kilos.kilos_bolsa like ? order by venta.fecha DESC LIMIT ?, 20;`, [busquedaMod, busquedaMod, busquedaMod, busquedaMod, salto]);
         return resultado;
     }
 
@@ -598,11 +619,37 @@ async function getUltimoClienteRegistrado() {
 }
 
 
+async function restarPuntosClienteById(idCliente, puntosParaRestar, puntosActuales) {
+    const conn = await getConnection();
+    await conn.query('update clientes set puntos = (? - ?) where id_cliente = ?', [puntosActuales, puntosParaRestar, idCliente]);
+    return (parseInt(puntosActuales) - parseInt(puntosParaRestar));
+}
+
+async function sumarPuntosClienteById(idCliente, puntosParaSumar, puntosActuales) {
+    const conn = await getConnection();
+    await conn.query('update clientes set puntos = (? + ?) where id_cliente = ?', [puntosActuales, puntosParaSumar, idCliente]);
+    return (parseInt(puntosActuales) + parseInt(puntosParaSumar));
+}
+
+
 async function getBolsasMain() {
     const conn = await getConnection();
     let resultado = await conn.query('select marca_bolsa from bolsas');
     resultado = transformarArrayDeRowDataPacketAStrings(resultado);
     return resultado;
+}
+
+async function getVentasActivasMascotaClienteById(idCliente) {
+    const conn = await getConnection();
+    let ventasActivas = await conn.query('select * from venta join bolsas_kilos on bolsas_kilos.id_bolsa_kilo = venta.id_bolsa_kilo inner join bolsas on bolsas.id_bolsa = bolsas_kilos.id_bolsa where venta.id_cliente = ? and venta.activo = true', idCliente);
+
+    for (let i = 0; i < ventasActivas.length; i++) {
+        const ventaActiva = ventasActivas[i];
+        let mascotasVentaActiva = await conn.query('select * from venta_mascota inner join mascotas on mascotas.id_mascota = venta_mascota.id_mascota where venta_mascota.id_venta = ?', ventaActiva.id_venta);
+        ventaActiva.mascotas = mascotasVentaActiva;
+    }
+
+    return ventasActivas;
 }
 
 
@@ -650,9 +697,10 @@ async function get18BolsasSegunBusquedaMain(newBusqueda, salto) {
     }
 
 
-    newBusqueda = newBusqueda + "%";
+    let newBusqueda1 = newBusqueda + "%";
+    let newBusqueda2 = "%" + newBusqueda + "%"
     console.log("newBusqueda: ", newBusqueda);
-    let resultado = await conn.query('select * from bolsas where bolsas.marca_bolsa like ? LIMIT ?, 18;',[newBusqueda, salto]);
+    let resultado = await conn.query('select * from bolsas where bolsas.marca_bolsa like ? or bolsas.calidad_bolsa like ? LIMIT ?, 18;', [newBusqueda2, newBusqueda1, salto]);
     return resultado;
 
 
@@ -674,18 +722,26 @@ async function getKilosBolsaPorId(id_bolsa) {
 async function actualizarDatosBolsaMain(newBolsa, bolsaKilos) {
     const conn = await getConnection();
 
-    newBolsa.marca_bolsa = (newBolsa.marca_bolsa).toUpperCase();
+    let bolsasConMismoNombre = await conn.query('select marca_bolsa from bolsas where marca_bolsa = ?', newBolsa.marca_bolsa);
 
-    await conn.query('UPDATE bolsas SET calidad_bolsa = ?, marca_bolsa = ? WHERE id_bolsa = ?;', [newBolsa.calidad_bolsa, newBolsa.marca_bolsa, newBolsa.id_bolsa]);
+    if (bolsasConMismoNombre.length == 0) {
 
-    await conn.query('delete from bolsas_kilos where bolsas_kilos.id_bolsa = ?', newBolsa.id_bolsa);
+        newBolsa.marca_bolsa = (newBolsa.marca_bolsa).toUpperCase();
 
-    for (let i = 0; i < bolsaKilos.length; i++) {
-        const element = bolsaKilos[i];
-        console.log("main:", element);
-        
-        await conn.query('insert into bolsas_kilos(kilos_bolsa, id_bolsa) values(?, ?)', [element, newBolsa.id_bolsa]);
+        await conn.query('UPDATE bolsas SET calidad_bolsa = ?, marca_bolsa = ? WHERE id_bolsa = ?;', [newBolsa.calidad_bolsa, newBolsa.marca_bolsa, newBolsa.id_bolsa]);
 
+        await conn.query('delete from bolsas_kilos where bolsas_kilos.id_bolsa = ?', newBolsa.id_bolsa);
+
+        for (let i = 0; i < bolsaKilos.length; i++) {
+            const element = bolsaKilos[i];
+            console.log("main:", element);
+
+            await conn.query('insert into bolsas_kilos(kilos_bolsa, id_bolsa) values(?, ?)', [element, newBolsa.id_bolsa]);
+
+        }
+        return
+    } else {
+        return "bolsaRepetida";
     }
 }
 
@@ -704,16 +760,59 @@ async function borrarBolsaByIdMain(id_bolsa) {
 async function crearNuevaBolsaMain(newBolsa, newBolsaKilos) {
     const conn = await getConnection();
     newBolsa.marca_bolsa = (newBolsa.marca_bolsa).toUpperCase();
-    await conn.query('insert into bolsas(marca_bolsa, calidad_bolsa) values(?, ?); ', [newBolsa.marca_bolsa, newBolsa.calidad_bolsa]);
 
-    newBolsa = await conn.query('select * from bolsas where marca_bolsa = ?', newBolsa.marca_bolsa);
+    let bolsasConMismoNombre = await conn.query('select marca_bolsa from bolsas where marca_bolsa = ?', newBolsa.marca_bolsa);
 
-    for (let i = 0; i < newBolsaKilos.length; i++) {
-        const element = newBolsaKilos[i];
-        console.log("main:", element);
-        
-        await conn.query('insert into bolsas_kilos(kilos_bolsa, id_bolsa) values(?, ?)', [element, newBolsa[0].id_bolsa]);
+    if (bolsasConMismoNombre.length == 0) {
+        await conn.query('insert into bolsas(marca_bolsa, calidad_bolsa) values(?, ?); ', [newBolsa.marca_bolsa, newBolsa.calidad_bolsa]);
 
+        newBolsa = await conn.query('select * from bolsas where marca_bolsa = ?', newBolsa.marca_bolsa);
+
+        for (let i = 0; i < newBolsaKilos.length; i++) {
+            const element = newBolsaKilos[i];
+            console.log("main:", element);
+
+            await conn.query('insert into bolsas_kilos(kilos_bolsa, id_bolsa) values(?, ?)', [element, newBolsa[0].id_bolsa]);
+
+        }
+    } else {
+        return "bolsaRepetida";
+    }
+
+
+}
+
+
+
+async function get20VentasPorBolsaSegunFiltros(filtro, filtro2) {
+    const conn = await getConnection();
+
+    console.log(filtro, filtro2);
+
+    if (filtro == "anio") {
+        anio = new Date();
+        anio = anio.getFullYear();
+        const resultado = await conn.query('select bolsas.marca_bolsa, count(venta.id_venta) as cantventas from venta inner join bolsas_kilos on bolsas_kilos.id_bolsa_kilo = venta.id_bolsa_kilo inner join bolsas on bolsas.id_bolsa = bolsas_kilos.id_bolsa  where year(venta.fecha) = ? group by bolsas.marca_bolsa order by cantventas DESC', anio);
+        return resultado;
+    } else if (filtro == "total") {
+        const resultado = await conn.query('select bolsas.marca_bolsa, count(venta.id_venta) as cantventas from venta inner join bolsas_kilos on bolsas_kilos.id_bolsa_kilo = venta.id_bolsa_kilo inner join bolsas on bolsas.id_bolsa = bolsas_kilos.id_bolsa group by bolsas.marca_bolsa order by cantventas DESC');
+        return resultado;
+    } else if (filtro == "elegir" && filtro2 != "") {
+        anio = filtro2[0] + filtro2[1] + filtro2[2] + filtro2[3];
+        mes = filtro2[5] + filtro2[6];
+        console.log("anio:", anio, "Mes:", mes);
+        const resultado = await conn.query('select bolsas.marca_bolsa, count(venta.id_venta) as cantventas from venta inner join bolsas_kilos on bolsas_kilos.id_bolsa_kilo = venta.id_bolsa_kilo inner join bolsas on bolsas.id_bolsa = bolsas_kilos.id_bolsa where YEAR(venta.fecha) = ? and MONTH(venta.fecha) = ? group by bolsas.marca_bolsa order by cantventas DESC', [anio, mes]);
+        console.log("resultado:", resultado);
+        return resultado;
+    } else if (filtro == "anioatras") {
+        let fechaActual = new Date();
+        let aniomenos = fechaActual.getFullYear() - 1;
+        let mes = fechaActual.getMonth() + 1;
+        console.log("mes que queremos ver:", mes);
+        let dia = fechaActual.getDate();
+        const resultado = await conn.query('select bolsas.marca_bolsa, count(venta.id_venta) as cantventas from venta inner join bolsas_kilos on bolsas_kilos.id_bolsa_kilo = venta.id_bolsa_kilo inner join bolsas on bolsas.id_bolsa = bolsas_kilos.id_bolsa where venta.fecha >= "?-?-?" group by bolsas.marca_bolsa order by cantventas DESC', [aniomenos, mes, dia]);
+        console.log("resultado:", resultado);
+        return resultado;
     }
 }
 
@@ -724,20 +823,20 @@ function transformKilosBolsaToFloats(rowData) {
     console.log("aaa", rowData);
     // Creamos un nuevo array donde guardaremos los datos convertidos a float
     const floatsArray = [];
-  
+
     // Recorremos cada objeto del array de row data
     for (let i = 0; i < rowData.length; i++) {
-      // Obtenemos el valor del único atributo del objeto y lo convertimos a float
-      const floatValue = parseFloat(rowData[i].kilos_bolsa);
-  
-      // Agregamos el valor al array de floats
-      floatsArray.push(floatValue);
+        // Obtenemos el valor del único atributo del objeto y lo convertimos a float
+        const floatValue = parseFloat(rowData[i].kilos_bolsa);
+
+        // Agregamos el valor al array de floats
+        floatsArray.push(floatValue);
     }
-  
+
     // Devolvemos el array de floats
-    console.log("aaaaa",floatsArray);
+    console.log("aaaaa", floatsArray);
     return floatsArray;
-  }
+}
 
 
 
@@ -757,6 +856,7 @@ let windowEditarCliente;
 let windowAgregarCliente;
 let windowEditarBolsa;
 let windowAgregarBolsa;
+let windowVentasBolsas;
 
 
 
@@ -768,6 +868,7 @@ function createWindow() {
     window = new BrowserWindow({
         width: 1100,
         height: 820,
+        autoHideMenuBar: false,
         icon: __dirname + './imagenes/favicon.png',
         webPreferences: {
             nodeIntegration: true
@@ -780,6 +881,16 @@ function createWindowEditarCliente() {
 
 
     cerrarVentanasEmergentes();
+    windows = BrowserWindow.getAllWindows();
+    console.log(`Hay ${windows.length} ventanas abiertas.`);
+    console.log(windows[0].BrowserWindow);
+    mainDocument = windows[0].webContents;
+
+    console.log(mainDocument);
+
+    // Establecer la propiedad opacity del body de la ventana principal
+    mainDocument.executeJavaScript(`document.body.style.opacity = ${0.1};`);
+    mainDocument.executeJavaScript(`document.body.style.pointerEvents = 'none';`);
 
     windowEditarCliente = new BrowserWindow({
         width: 900,
@@ -799,6 +910,16 @@ function createWindowEditarCliente() {
 function createWindowAgregarCliente() {
 
     cerrarVentanasEmergentes();
+    windows = BrowserWindow.getAllWindows();
+    console.log(`Hay ${windows.length} ventanas abiertas.`);
+    console.log(windows[0].BrowserWindow);
+    mainDocument = windows[0].webContents;
+
+    console.log(mainDocument);
+
+    // Establecer la propiedad opacity del body de la ventana principal
+    mainDocument.executeJavaScript(`document.body.style.opacity = ${0.1};`);
+    mainDocument.executeJavaScript(`document.body.style.pointerEvents = 'none';`);
 
 
     windowAgregarCliente = new BrowserWindow({
@@ -814,18 +935,28 @@ function createWindowAgregarCliente() {
     })
     windowAgregarCliente.loadFile('src/ui/agregarCliente/agregarCliente.html'); //indica el archivo que se cargara en la ventana
 }
-
-
+let windows;
+let mainDocument;
 function createWindowEditarBolsa() {
 
-    cerrarVentanasEmergentes()
+    cerrarVentanasEmergentes();
+    windows = BrowserWindow.getAllWindows();
+    console.log(`Hay ${windows.length} ventanas abiertas.`);
+    console.log(windows[0].BrowserWindow);
+    mainDocument = windows[0].webContents;
+
+    console.log(mainDocument);
+
+    // Establecer la propiedad opacity del body de la ventana principal
+    mainDocument.executeJavaScript(`document.body.style.opacity = ${0.1};`);
+    mainDocument.executeJavaScript(`document.body.style.pointerEvents = 'none';`);
 
     windowEditarBolsa = new BrowserWindow({
         width: 550,
         height: 450,
         alwaysOnTop: true,
         frame: false,
-        resizable: true,
+        resizable: false,
         icon: __dirname + './imagenes/favicon.png',
         webPreferences: {
             nodeIntegration: true
@@ -838,13 +969,23 @@ function createWindowEditarBolsa() {
 function createWindowAgregarBolsa() {
 
     cerrarVentanasEmergentes();
+    windows = BrowserWindow.getAllWindows();
+    console.log(`Hay ${windows.length} ventanas abiertas.`);
+    console.log(windows[0].BrowserWindow);
+    mainDocument = windows[0].webContents;
+
+    console.log(mainDocument);
+
+    // Establecer la propiedad opacity del body de la ventana principal
+    mainDocument.executeJavaScript(`document.body.style.opacity = ${0.1};`);
+    mainDocument.executeJavaScript(`document.body.style.pointerEvents = 'none';`);
 
     windowAgregarBolsa = new BrowserWindow({
         width: 550,
         height: 450,
         alwaysOnTop: true,
         frame: false,
-        resizable: true,
+        resizable: false,
         icon: __dirname + './imagenes/favicon.png',
         webPreferences: {
             nodeIntegration: true
@@ -853,9 +994,45 @@ function createWindowAgregarBolsa() {
     windowAgregarBolsa.loadFile('src/ui/agregarBolsa/agregarBolsa.html'); //indica el archivo que se cargara en la ventana
 }
 
+function createWindowVentasBolsas() {
+
+    cerrarVentanasEmergentes();
+    windows = BrowserWindow.getAllWindows();
+    console.log(`Hay ${windows.length} ventanas abiertas.`);
+    console.log(windows[0].BrowserWindow);
+    mainDocument = windows[0].webContents;
+
+    console.log(mainDocument);
+
+    // Establecer la propiedad opacity del body de la ventana principal
+    mainDocument.executeJavaScript(`document.body.style.opacity = ${0.1};`);
+    mainDocument.executeJavaScript(`document.body.style.pointerEvents = 'none';`);
+
+    windowVentasBolsas = new BrowserWindow({
+        width: 900,
+        height: 650,
+        alwaysOnTop: true,
+        frame: true,
+        resizable: false,
+        icon: __dirname + './imagenes/favicon.png',
+        webPreferences: {
+            nodeIntegration: true
+        }
+    })
+    windowVentasBolsas.loadFile('src/ui/ventasBolsas/ventasBolsas.html'); //indica el archivo que se cargara en la ventana
+}
 
 
 function cerrarVentanasEmergentes() {
+
+    try {
+        mainDocument.executeJavaScript(`document.body.style.opacity = ${1.00};`);
+        mainDocument.executeJavaScript(`document.body.style.pointerEvents = 'auto';`);
+    } catch (error) {
+        //no hace nada
+    }
+
+
     if (windowAgregarCliente != undefined) {
         try {
             windowAgregarCliente.close();
@@ -880,6 +1057,13 @@ function cerrarVentanasEmergentes() {
     if (windowAgregarBolsa != undefined) {
         try {
             windowAgregarBolsa.close();
+        } catch (e) {
+            //no hace nada
+        }
+    }
+    if (windowVentasBolsas != undefined) {
+        try {
+            windowVentasBolsas.close();
         } catch (e) {
             //no hace nada
         }
@@ -915,7 +1099,6 @@ function restarDias(fecha, dias) {
     // Crear un nuevo objeto Date con el resultado y devolverlo
     return new Date(resultado);
 }
-
 
 //exporta la funcion para poder ser usada en otro lado
 module.exports = {
@@ -970,5 +1153,11 @@ module.exports = {
     getSoloBolsaByIdMain,
     borrarBolsaByIdMain,
     createWindowAgregarBolsa,
-    crearNuevaBolsaMain
+    crearNuevaBolsaMain,
+    restarPuntosClientePorBorradoDeVenta,
+    restarPuntosClienteById,
+    sumarPuntosClienteById,
+    getVentasActivasMascotaClienteById,
+    get20VentasPorBolsaSegunFiltros,
+    createWindowVentasBolsas,
 }
